Update experimental-procedure.md
Professor @kelloggm, here is the pull request to change the ```gradle``` command into ```./gradlew```. 

I also think that the term ```human-written file``` seems to be vague, as there is a folder named ```human-written```. Since what we need here is not the absolute path of the folder ```human-written``` but the absolute path of the .java file inside that folder, I think it will be clearer to change ```human-written file``` into  ```.java file```